backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
black==25.1.0
click==8.2.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
mypy==1.15.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.5
platformdirs==4.3.8
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
